requirements_docker.txt
mysql-connector
mysql-connector-python
pandas
os
sys
numpy
pytest
scipy
boto3

